Fixed JWT CVE, saved token for extra security
Removed the hapi-auth-jwt plugin in favor or writing our own plugin that
also returns the token to the server. This allows the server to save the
token to memDB which also allows the server to track user sessions. All
auth methods have been updated to utilize the new plugin and token
return.