Updated Editor look and feel
Updated the editor's look and feel by removing the post title which is
not shown on the posts page anyway. The input space is now much larger
and feels like much more could be written there. The image uploader was
also updated with centered text which is also more appealing.

The editor input itself has switched from a contenteditable div to a
textarea. This greatly simplified a lot of the code. As so, the posts
controller and html file was updated for the above changes.

The CSS file was also changed due to the nature of the above task.