Refactor image uploader look and feel
The image uploader now floats without a border unless hovered over for a
cleaner look. The CSS has been updated to acheive this effect and the
text in the image-uploader directive has been shorten to allow the modal
link to fit in the image upload container. The change in color on hover
has also been updated to fit both the new thread page and the editor
page as well.

On the new thread page, the title has been moved outside the form as
this is how the editor is now composed. Doing so meant updated the
disabling of the submit button.