Double login fix
Fixed an issue where using the enter button on the login modal would
cause the login route to be called twice. Basically angular and html
forms will automatically trigger the function attached to a button in a
form, if and only if that button is the only button in the form.
Therefor the ng-keyup was removed and the validation check within that
ng-keyup function was moved to the login function on the controller.