Refactor Auth system
The Authentication system on the client side has been refactored. The
user's session information has been pulled out into it's own service,
session. The User routes that were routed through the Auth service has
been removed. All other files that relied on those methods have been
updated to reflect this change.

The main reason to pull out the user's session was so that the session
code is only in one place being so, the rest of the code has been
updated. The reason why it can't all stay in the Auth service is because
doing so causes a circular dependency with the AuthInterceptor.