Refactored modal
Updated the modal directive to support the auto-focus directive within
the modal directive and to support the cleaning of forms on modal
closing. This mean updated all the modal in the project to better
conform to how the modal directive is currently laid out. The biggest
change is that all forms with in the modal should be named with
'$parent.form' which binds the form to the modal directive. This allows
the modal to clean the form of any angular generated errors or invalid
styles.

The auto-focus directive found within a modal so also be bound to
'$parent.focus' so that the modal can re-trigger the auto-focus
directive on modal open.

All the modals throughout the project making use of the modal directive
have all been updated to follow the new way of dealing with modals. The
header and profile page has also been refactored for readability and
simpler variables.