Remove page reload, profile editable moved to client side
The check to see if a profile is editable has been moved to the client
side rather than doing the check on the server side. This is
accomplished by checking if the id of the profile being viewed matches
the id of the currently logged in user. If so, a editable variable is
set on the profile controller which allows the user to access the
controls that allow the user to edit the profile.

This also allows the client side to not rely on reloading the page when
logging in and loggin out. Because the all the user related controls can
now refer to the session service, nothing needs to be reloaded.