New Alert service
The new alert service is a system wide module that allows any other
module to send the user an alert. This alert can be in the form of
success, info, error, and warning. Each alert has a different background
color and is laid out on top of all other content in a fixed position.

The interface to this module is to include the module in the dependency
injection and to then call the module using one of the methods mentioned
above. For example, Alert.success('message') will send the user an alert
with a green background color with the text content of 'message'.

This commit also replaces all the error boxes in all the modules with
the new Alert system and any success messages as well.