Optimized reset, confirm pages, added back reload on log in/out
The reset and confirm pages have been optimized to use promises over
callbacks as well as removing unnecessary variables and statements.

In the login and logout method in the header controller, the line of
code that reloads the current controller has been added back in due to
the nature of our application's settings. Some settings required the
content to be pulled down again, loginRequired, admin pages, and profile
page for example.