Image upload progress bar, error handling
The image uploader was updated to include a progress bar while uploading
images to give the user some feedback as to what is happening. Before,
it would give no feedback until the image was done uploading. Also
errors were not being properly handled, it is still not properly handled
but at least a error is being shown to the user. The cause of the error
is not displayed to the user yet.

The error handling system will need to be refactored more so that error
objects can be consumed and proper errors can be shown to ther user.