CSS-Tooltips on posts page
Created a CSS class called css-tooltips that creates a tooltip when
hovered over the element with that class. The text in the tooltip comes
from the data-css-tooltip attr on the same element that holds the
css-tooltip class. This tooltip should work for most use cases but
another class to set a defined width could be added later for edge
cases.

The tooltip has been added to all the icon in the posts page.