Thread Sticky
Implemented thread sticky on the frontend side of the code. This
includes all the UI tweaks needed to handle sticky threads being
returned along with normal threads from the byBoard method. This also
handles all the work needed to allow admins to sticky a thread on thread
creation or after it is created.