Updated category editor to match board-mapping in core-pg
Updated the category editor to match all the major changed in core-pg in
the board-mapping branch. The major change is the removal of the
parent_board_id, children_ids, and category_id property from all boards.
Also, the backend does not need to be aware of boards being moved
anymore so that entire section was deleted.