Updated category editor to output board mapping array
Updated the category editor to output an array of entries that represent
the hierarchy of categories/boards structure. This is a much more
efficient way of handling a nested tree structure.

The way categories are created have also been modified. Like boards, the
categories are created before the tree is parsed into an array. This is
to ensure that all the category and board ids are accounted for before
they are are turned to array entries and shipped off to the
updateCategories method.