Post Deleteion routes and pre
Updated the routes to include two new routes to handle undelete and
purging of post. The route to delete a post already existed but was not
being used. Now, all three routes successfully soft delete and
permanently delete posts as labeled.

The majority of the work was done in the posts/pre.js file where the
permission handling was reworked to return an error if any of the
pre-conditions are not met for that route. This will be a new pattern
that should be adopted by the other routes in the future but may be
subject to change as well.

The work above includes moving any pre handlers that checked if the user
is an admin, owned the post, or any other such conditions into one
pre-condition handler that is expected to join all the other ones
together though one promise that can be returned to hapi. If any of the
preconditions fail, an error is returned to hapi and the route is
denied. This also involves moving the commonPre out of the config file
and is only linked to the pre file of each route model.

On the client side, three new button were added to the post headder
section that allow the appropriate user to purge, delete and undelete
posts with regards to their roles and post ownership. Each of those
buttons lead to a confirmation modal that prompts the user to ensure
they are taking the right action.