Updated posts page to use thread object
Previously the posts page pulled in some of the thread properties as
individual propreties. This commit changes it so that the entire thread
object is pulled in as one property. This was done because a majority of
the properties were being pulled in and made more sense to just pull in
the thread object itself.

The html and controller files were all updated to reflect this change in
properties.