Thread Deletion routes, pre, and angularJS
Update the routes to include three new routes to handle delete,
undelete, and purging of threads.

The majority of the work was done in the threads/pre.js file whre the
permission handling was reworked to return an error if any of the
pre-conditions are not met for that route. This is the continuation of a
pattern that was introduced with Post deletion.

The above work includes moving any pre-handlers that checked if the user
is an admin, owned the thread, or any other such conditions into one
pre-condition handler that is expected to join all the other ones
together through one promise that can be returned to hapi. If any of the
pre-conditions fail, an error is returned to hapi and the route is
denied. This also involves moving the commonPre out of the config file
and into the pre file instead.

On the client side, three new buttons were added to the thread controls
section that allow the appropriate user to purge, delete, and undelete
threads with regards to their roles. Each of those buttons lead to a
confirmation modal that prompts the user to ensure they are taking the
right action.