Bug Fix for post paging
Fixed a bug where post paging and updating the screen when a new post is
created fails with a 400 error. This was due to the thread_id not being
sent correctly to the server. This issue stemmed from the commit where
the thread properties were condensed into one object shared with the
PostsParentController and having missed a few key lines of code.