Board permissions, Board deletion from management page
On the client side, the management page has been updated with a trash
icon that sits on each board, allowing the user to delete a board. This
is a hard delete of a board and all of it's children content minus child
boards. Child boards are sent off to the 'Uncategorized Boards' section
to the right of the page. All threads and post for the deleted board are
also deleted. A new delete confirmation modal was created for the board
deletion while the existing modal was converted in a modal for category
deletion but has not been put into use yet.

On the server side, all the permissions across boards, thread, and posts
have been updated to respect which boards are on the front page and
which aren't. If a board is not on the front page, then that board
cannot be access by anyone outside of admins and moderators. Also,
various permissions have been tweaked such that admins and mod have near
full access to all server side methods regardless of other conditions
such as thread being locked, deleted, etc.