Category Deletion, client side
This is a companion commit to the commit on core-pg with the same commit
title. It allows the user to delete a category from the admin management
page. It also accounts for new categories created and not saved. This
means that new categories created and not saved, but deleted within the
same session are cancelled out and no calls to the server are made.