Bug Fix: Call to boards.all forbidden for non-admin
This fixes a bug created by a previous commit where the boards.all route
was made private to admins only. But the route is also made on every
view to a posts page. This call is made to supply the list of boards for
the thread moving action. This route should only be called if the viewer
is an admin. This commit fixes the issue by checking of the viewer is an
admin and only then makes the call to boards.all