User Active and thread action changes
Added in more permission checking for user activation for thread
actions. Mainly the thread creation and thread locking have been updated
to check if a user is active before allowing these actions to take
place.

Bug fixes to posts.pre are also included to fix the issue with regular
users not being able to create, edit, or delete posts.