user activation and delete controls, bug fixes
Updated the client side code to include the controls for users to
deactivate or reactivate their accounts. Also added an admin only
control to delete a user's account. The permissions needed to adjusted
slightly so that when an admin sends the command to
deactivate/reactivate/delete a user account, it is done on the user id
being sent in and not the admin themself. This required adding the id of
the user to modify in the route itself.