Private Messages
Implemented a page for private messages. This includes all the backend
work to create all the routes necessary to serve out the private
messages and conversations, create messages and conversations, and
delete those if needed. The encompassing work also includes all the
business logic to determine who can create new messages and who are
allowed to delete messages and conversations.

A new directive to handle checking usernames again user ids was also
created to help the user.

The view itself lists the conversation sorted by time on the left hand
side, while the user can choose to view messages which will appear on
the right hand side. A few other items where created to help the user
such as scrolling to the most recent message on loading a conversation.