Bug fix: wrap numerical value with Number()
Some of the config variables need to stay as numbers even if foreman or
another env tool loads the values as string. Wrapping those config vars
with Number() will ensure they stay as numbers. This does not solve the
edge case where a non numeric string is set as the value for those env
vars.