Bug Fix: limit update for report notes
The report notes were missed during the last commit. This effectively
ignores the limit and page options because the report notes do not have
a way to paginate. Also it is unlikely to have many views or many notes.

In the meantime, the limit query param has been removed from all
frontend calls while the validation on the backend has been updated in
case a way to paginate is one day created.