Bug Fix: admin/mod can post while thread locked, refactoring
Fixed a bug where the "+" button did not show up for admins/mods if the
thread was locked. The backend already handles allowing admins/mods to
post if a thread is locked. Now the frontend has been updated to do the
same.

The posts controllers and html was also refactored a bit to remove
unnecessary code and watchers. One time binding was used where
appropriate as well as simplifying some of the code.