Support for updating thread title
Added the ability to update the thread's title by admins, moderators,
and the thread owner. Upon clicking the pencil icon next to the thread
name, the appropriate user is able to change the title of the thread.

This required implementing a new route on the server side and a new
method in the db to handle updating the first post that is backing the
thread title.