Exposed a method to check permissions from pre/request
Added a method to the acls plugin that is also exposed to the rest of
the server through the plugin. This method takes in the request.auth and
the name of ther permission to check and returns whether that permission
is true for this particular user. It works generally the same as the
'onPostAuth' server extension used to check each route's acl but differs
by returning false if no acl is passed in.