Support for saving roles to DB from defaults
The code currently has a set of default roles that are hard coded. This
commit adds support to sync those hardcoded roles witht he DB table
'roles' if they are ever erased from the DB. The additional roles
entered into the table are also loaded into memory for the rest of the
system to use.