Nearly finished work on post pre
More work was done with the posts pre methods to complete the update to
using the permissions rather than hard coded logic. The only issue left
is finding a way to show the user's deleted posts on the user's profile
page if a moderator is visiting.

A new way to handle how to pass parameters to the pre method was also
initiated in this commit and should be used through the rest of the
project.