Added a permission to delete messages, style fixes
Added a new permission that allows those designated, to delete a message
from the PM system. Before the private message could only be deleted by
the person that sent the message. This allows all with permission, to
delete any message.

Some of the auth files were updated to use the latest coding styles in
relation to promises.