Implemented the unread link, code refactoring
Some of the user thread view code was refactored to be cleaner and more
readable. In particular, the code that checks to see if a user view is
valid has been updated to be smaller and more maintainable.

The threads.byBoard was also updated to push as much of the thread
viewed checking to the DB as possible. This should cut down the latency
within this method. With that, the front end was updated with links to
the very last post in a thread, and a link to the last unread post since
the user last visited that page.