Initial Watchlist implementation
This commit is the frist run at the implementation of a watchlist
feature that allows users to basically subscribe to a thread or board.
This subscription allows users to view only the threads they are
subscribed to ordered by the last time the thread has been posted to.
There is a specific page the users must navigate to to view they
'watchted' threads. There is also an 'Edit' page where the user can
update their subscriptions.

The resources, app, and config files were all updated to integrate with
the new html, controller, and routes created on the server side.