Initial plugin implementation
This is the first draft of the new plugin system. This may or may not be
the version used as the public plugin interface but for now, it will be
used for internal plugin structuring. The plugins are essentially npm
module that exposes one object. That object should have up to 4
properties: db, endpoints, routes, and templateHooks. There is also a
requirement where if the plugin is adding functionality to the front
end, a templates dir should also be included where those frontend file
should be.