Rate-Limiting Initial draft
This is the first draft of the rate-limiting system for epochtalk. The
rate limiter has three global limits one for each of: get, post, delete
HTTP methods.

On top of the global limits, individual limits can be placed on any
pairing of HTTP method and route path (GET /api/boards) and then placed
on any specific role (users). This allows the global limits to rate
limit all routes but allows the admin to specific more directly limits
on a group of users (newbies?).