Updated rate limits to include put, web ui
Updated the rate limits to include a global limit on PUT calls to the
server. All other types of calls are still defaulted to the POST
configs.

Added a web ui to both the admin settings for global rate limits and to
the role's modal for role based rate limiting.