Updated the slide-toggle, poll permissions
Updated the slide-toggle directive to be CSS based rather than JS based.
This includes updating the board's version of slide-toggle and updating
the poll edit view's slide-toggle so that they are both the same.

The poll edit permissions was also updated to be a property on the
permissions object rather than being computed at the view html.