Bug Fix: Editing posts changes title
This is a fix to a bug where if a user edited the first post in a
thread, it would change the thread title to 'Re: ' + original thread
title. The fix will preserve the original post title if it exists and
assign 'Re: ' + thread title if it does not exists.