Permanent fix for user confirm
This is a permanent fix to the bug that would 500 the user confirm
route. The hotfix would attach an empty roles array to the user object
before sending it to the helper.saveSession method. This commit fixes
the problem by returning the full user object to the method prior to the
helper.saveSession call which already has a roles property attached to
it.The upside of this fix is that an admin can assign a role to the user
before they have confirmed their account.