Bug Fix: thread page/limit changed by query
This is a fix to a bug that would change the page/limit of the board
view after returning to the board view via the back button. The limit
that is returned is set to a erroneous value by the server and used on
the front end displaying the wrong number of pages.