Private boards access
Implemented the ability to hide boards via role priority. A new column
was added to both boards and categories that allow setting the role
priority needed to access that model. The access method ensures that the
user role priority is less than or equal to the model's priority.

The admin UI was updated to also set this priority number on the model.