Private board admin UI
Updated the management boards page to allow updating of the board and
category minimum view level. This allows the admin to set who can see
the boards or cats based on their role. Currently the system will roll
up user priority levels. So a board set with 'moderator' as the minimum
view level is also viewable to admins and super admins.