Keep user logged in
This commit updates the frontend to use only localStorage to keep user
information. The benefit of this is that users stay logged in even when
opening a new tab.

On the server side, the rememberMe boolean is passed to the backend so
that the expiration on the token can be set. By default, a user is kept
logged in for two weeks unless they select rememberMe.