Bug fix: slide toggle clipped content
The slide toggle directive was clipping content on page resizes. This
fixes the problem by using a bit more JS to set the max-height property
to inherit after the animation is complete. This also fixes a bug where
the height of the content needed to always be saved. Now the height is
calculated by added an extra 'slide-wrapper' div within the slide-toggle
div to keep track of the height for us.