Recent Posted In
Created a new routes, views, and permissions for viewing recent threads
the viewer has posted in. This takes advantage of a new db method
created in the sister branch in core-pg. This page is accessible through
a link on the front page, in the recent threads section or through the
url '/threads/posted'

This commit also refactors some of the angular code so that each module
or view is in their own folders rather than being lumped into the same
folder.