Modularized sanitizer, promisified server, prep images
The sanitizer file has been turned into a hapi plugin and decorated on
server and request.

The server/index.js file was promisified thanks to the latest hapi
update that allows registration calls to return a promise.

The images folder and all it's references has been prepped so that it
can be converted into a hapi plugin as well.