Module infrastructure
This commit contains all the works needed for a preliminary module
system that allows pulling in extra modules to be added on to the core
hapi server instance.

The infrastructure includes adding a task in the build scripts that will
call `npm install` in the /modules subdir and copy all the client side
files into the /app/modules subdir.

On Hapi, a plugin was added that will load each dep installed through
the previous `npm install` and load all the auth, common, routes, perms,
and db methods to the correct hapi plugins.

Webpack was also updated to create a new module context and resource
context that will pick up all the files copied to /app/modules and load
them into angular.

All deps are defined in the package.json file in /modules.