Modular System update
This update sets in stone the way modules will be handled in epochtalk.
This includes changes to how the server's global components are
structured; authorization, common methods, acls, imageStore and routes.
Both the frontend code and the bootstrap scripts were also updated to
account for how modules will be loaded in the future.

This first release has already removed the posts module and can now be
found in ept-posts or bct-posts.