Bug Fix: multiple board updates
This fixes a bug where when an admin tries to update the name and/or
description of multiple boards, only the first one is processed while
the rest are ignored.

This fixes the bug by sending the list of changes as an array rather
than calling update on each item in the array.