Bug Fix: 429 Error on image upload (local)
Fixed a bug where uploading multiple images using the local image store
caused a 429 error. The fix involved updating the policy route to handle
an array of filenames rather than one filename at a time. Also, the
local upload route was updated with no minDifference in the limiter but
only allows 10 uploads per second at the most.

A smaller bug was also fixed that appeared when switching from local to
S3. The S3 client was not re-initialized when the settings were changed.
Now an explicit call to the global image store code is made and that
re-initializes both local and S3 versions.