Added file watching for frontend and backend code
These watchers will watch the files in /modules/node_module filtered by
the given project names. This is to ensure that if work in done in other
repos, as long as they are symlinked to this repo, the changes will
propagate up and will cause either nodemon to refresh or have the
changed files copied to the app dir. This only happens if the scripts
are called while in a development environment.