Hooks Infrastructure
Establishes all the infrastructure needed to support pre and post
processing function in the routes that are configured to do so. The main
body of work lies in the module system picking up any pre/post methods
and appending it to a hooks object that lives on the request object at
request.hooks.

The functions themselves has very specific contract agreements. The
first being that they are called with only the request object as a
parameter. The second being that the pre methods must return one object
with a path property and a data property. The path property designates
where to append the data object on the final payload. This does require
knowing what the final payload object will look like. The last agreement
is that the post methods do not need to specify where their data is
going to be place but rather to do it themselves because they will have
access to the final payload at 'request.pre.processed'. They may replace
or append any data they need to at this point.