Favicon
Update how the favicon is set by the config and returned to the client.
Instead of binding the favico so a link tag using angular, we'll let the
browser naturally grab the favicon at the default location but use hapi
to intercept that call and return the admin defined favicon. If there is
not favicon defined, a default one has been created and saved to the
usual location.