Modularize Auth, emailer plugin
Modularized the authentication based routes into its own module. This
required moving all the auth methods into anotehr repo and the single
authorization method.

The emailer needed to be turned into a plugin and decorated on the
server and request objects for hapi. This also fixed a bug where if the
emailer settings were changed in the admin view, the emailer wasn't
updated. The email settings and the transport now re-initialize when the
settings are changed.