Updated the hooks to include a process before processing
The original structure of the hooks design did not include a way to run
code before the main processing occured. This commit adds an area to
process code before the main handler, called pre. The original pre is
now labelled parallel because it runs in parallel to the main handler.
PostProcessing is left alone.