Modularize parsers
This removes the epochtalk-bbcode-parser dep in favor of moving the
parser compiler into a module that can be dynamically loaded rather than
being a default lib included with this repo. This has a side effect of
being able to stack parser if wanted. Support for a modularized parser
has been added to the modules plugin and the parser plugin.