Fixed S3 config issues
On a fresh instance, the default configuration was not sanitized and
lead to a bad object key while creating the images/ folder in the
bucket. The default configurations were fixed as well as making sure the
configs are always sanitized during initial startup, any further
startups, and after settings are changed in the /admin/settings page.