Bug Fix: Multiple session logout
This commit fixes a bug where if a user signed out of one machine, the
server would log them out of all other machines. The reason being that
the wrong key was being referenced to check if the user still had
sessions on other machines. Because the server saw that no other session
existed, the moderating and roles keys for this user was also deleted.
The session keys for other session still existed but because the other
information couldn't be pulled, the server assumed the user had no other
sessions. Thus, if the user used another machine the server would assume
that session was dead also.

The error was a simple mispelling of a key. This have been fixed.