Backoff policy
Created a new plugin called backoff that tracks api usage on marked
routes and only allows for two calls to the route before adding a cool
down period that exponentially grow as the api is access while in cool
down. The max cool down period is 65536 minutes or roughly 44 days. At
which point the user is allow one chance to access the route in a
non-abusive manner or another max cool down period is added for that ip.

There is also a release value that will clear all cool down periods for
a route when the correct behaivor is acheieved.

The quick fix for /api/recover in the limiter plugin has also been
removed in favor of this plugin.