Add support for Google Analytics key on page load
Adds the Google Analytics script at the bottom of the public/index.html
apge and the public/recover.html pages. This allows GA to be loaded on
every page view. The GA key itself is pulled from env vars and loaded
into the html pages through handlebars and hapi. If a key is not given
in env vars, the default 'UA-XXXXX-Y' key is used which is typically
ignored by Google.