Footer Copyright & Legal page
Added a legal page that contains some default text for a terms of
service, privacy policy, and disclaimers. The default text is stored in
a new folder called defaults while any modified text is stored in the
legal dir of the 'content' dir. The legal page itself is a static page
generated in the '/legal' route that pulls a copy of either the modified
text, if it exists, or the default text. The route itself is defined in
the server/routes/index.js file.

There is also an admin legal page that allows the superAdmin to modify
the legal text and save it or reset it to the default text. This view
required three routes: text, reset, and update. Which, are defined in
the server/routes/admin/legal dir.

These routes are guarded using the adminLegal acls that were added tot
he server/plugins/acls/roles.js file. An additional permission:
adminAccess.settings.legal was also added to guard the page that edits
the legal text.

Additionally, the footer has been pulled into the public/index.html
page. The angular module for footer has been removed. The static footer
dropped any dead links and now only contains links to different sections
of the legal page as well as a server defined copyright. The copyright
is a mash up of the current year and the forum title.