Substituted Avatar urls
The avatar urls are substituted with either a local url or an s3 url
instead of just using the hotlinked address. This only pretains to urls
that do not resolve to the server's domain (hotlinked images).

This changes takes effect as a add on to the imageStore plugin. A new
hapi server method was added to substitute the original hotlink url with
a local version.