Add support for the portal system
This commit adds support for the portal system by adding it to the
modules/package.json file, updates to the configurations system for
added system variables, and an update to the public/index.html route to
push the 'portal.enabled' var to the frontend code.

Updates were also made to the authorization and settings routes to allow
for passing the portal variables in and out.