0.4.2: Invitations
Updated to support the invitaitons system.

The inviteOnly config var was added to both the config.js and
configurations.json file. The setup file was also updated to handle the
new var. The settings route was also updated to handle passing the
invite_only var through to the frontend and back.

The roles were also updated to handle the new invitations view in
admin/management. The emailer was also updated to include a new email
template that's sent out when a new user is invited.

Lastly, the package.json file in /modules was also updated with the
newest version for each package affect by the invitations system.