0.4.5: Saas Mode
Updated the configurations to allow for a saasMode var that tracks
whether this server instance is running in SAAS MODE or not. This var is
passed to the admin frontend on admin/settings/general.

The settings.find route also now passes the in memory config to the
frontend instead of the db version. Although any saves are made to both
the in memory config and the db version.

While is saas mode, the sanitizeConfig function is bypassed as the
emailer and images config are no longer populated.

All packages in the /modules folder is now up to date.