0.4.6: Bug Fix for saving ga_key, README update for SAAS_MODE
The readme file was updated with a quick blurb on how the SAAS_MODE
works.

A bug fix was made to the settings.find route. The ga_key property could
be null coming out of the db so it now defaults to an empty string. This
is important because the properties are modified in place by the
frontend and sent back to the settings.update route. The settings.update
route does not allow for undefined or null for any property.