0.4.7: user last active, private mode slash page
Added a hapi plugin, last_active, that watches every incoming connection
and checks to see if a auth user is available, if so, it'll check that
users' last active timestamp and if the time elasped in greater than one
minute, it'll update that value.

ept-frontend was also updated to 1.0.9 and ept-users was updated to
1.0.4.

ept-users was updated to support user last active
ept-frontend was also updated to support user last active but also
includes an update to the private mode which now displays a splash page
if the user is not logged in.