0.4.10: Bug Fix for portal.board_id as '' breaking saves
The general settings page could not be saved if the portal.board_id
value was a falsy value. The fix was to update the validation to no
longer care if portal.board_id is a falsy value but to defer that check
to another function in the pre. By adding '.allow('')' to the
portal.board_id validation, Joi stopped failing on portal.board_id being
an empty string.