0.4.12: Default Avatar, updated deps
A new default avatar image exists now in /public/img/avatar.png. This
should be used as a default avatar instead of relying on another service
like fakeimg.

ept-frontend updated to 1.0.12 to reflect the changes in the frontend
since a default image and url to that image is saved in the db.