sync default rate limit for GET
see 76f6629 and 862f9d7

to prevent rate-limiting of our own app's calls,
76f6629 changed the minDifference from 50 to 0

for testing purposes,
this branch changed the same value from 50 to 0, but the revert commit
862f9d7 reverted the value to 50...  after 76f6629 was merged into this
branch.  this recorded as a clean merge, since the value was 0 in both
branches, at that time.

thus, the real value should be 0.

* 862f9d7 "reverted" d832d64, changed 0 to 50
|
* (merged)
|\
...
| |
| * 76f6629 changed to 0 on purpose
| |
...
| |
* | d832d64 changed to 0 for testing
| |
|/
* some base commit...