handle empty (null) emailer config
after the initial run, where defaults are created, db calls to
configurations will bring back an object with emailer fields nulled out.

this configuration gets past the undefined check that works on the first
run.

the solution:  check for null emailer sender, and handle it the same way
an undefined config emailer object would be handled.