adjust source merit range for current source merit
when calculating the currently available sendable source merit,
the source merit range should be the latest one.  if the source merit
range needs to be updated, then it wasn't at the latest one.

in that case, drop the sent source merit range and reset its sum to 0

if the current source merit position is the latest one, then everything
in the source merit sends range is correct and should be included in the
calculation